change: remove duplication in FAQs
@vryser

Can you please validate this commit? It's a good idea to avoid
duplication of content: if we have the list of 'Depots' in two different
places, chances are that we forget to properly keep them in sync.
Instead, we can write it only on one place and link to it when needed.

In general I'd keep the FAQs short and linking back to the paragraph or
section of the webpage where the subject is discussed, rather than
writing it twice. Please feel free to rephrase the FAQs to remove
duplication in that sense.

Cheers.